refactor(rules): rename rules
Fix #17
