+++ src/index.ts
@@ -1,4 +1,4 @@
-export { Rule as CollapseRxjsImports } from './collapseRxjsImportsRule';
-export { Rule as UpdateRxjsImports } from './updateRxjsImportsRule';
-export { Rule as MigrateToPipeableOperators } from './migrateToPipeableOperatorsRule';
-export { Rule as MigrateStaticObservableMethods } from './migrateStaticObservableMethodsRule';
+export { Rule as RxjsCollapseImports } from './rxjsCollapseImportsRule';
+export { Rule as RxjsProperImports } from './rxjsProperImportsRule';
+export { Rule as RxjsPipeableOperatorsOnly } from './rxjsPipeableOperatorsOnlyRule';
+export { Rule as RxjsNoStaticObservableMethods } from './rxjsNoStaticObservableMethods';
